+++ src/main/java/roomescape/member/dto/MemberLoginCheckResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.dto;
 
 public record MemberLoginCheckResponse(
         String name
