+++ src/main/java/roomescape/reservation/ReservationDao.java
@@ -4,6 +4,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.reservation.dto.ReservationRequest;
 import roomescape.theme.Theme;
 import roomescape.time.Time;
